bevy_render: reorganize shape module
